Moved moderation and audit commands into proper packages
