+++ application/src/main/java/org/togetherjava/tjbot/features/moderation/audit/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * This packages offers all the functionality for auditing moderation actions, see
+ * {@link org.togetherjava.tjbot.features.moderation.audit.AuditCommand} as main entry point.
+ */
+@MethodsReturnNonnullByDefault
+@ParametersAreNonnullByDefault
+package org.togetherjava.tjbot.features.moderation.audit;
+
+import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;
