--- application/src/main/java/org/togetherjava/tjbot/moderation/package-info.java
@@ -1,10 +1,0 @@
-/**
- * This packages offers all the functionality revolving around moderation.
- */
-@MethodsReturnNonnullByDefault
-@ParametersAreNonnullByDefault
-package org.togetherjava.tjbot.moderation;
-
-import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;
