--- application/src/main/java/org/togetherjava/tjbot/routines/package-info.java
@@ -1,14 +1,0 @@
-/**
- * This package contains most routines of the bot. Routines can also be created in different
- * modules, if desired.
- * <p>
- * Routines are actions that are executed periodically on a schedule. They are added to the system
- * in {@link org.togetherjava.tjbot.features.Features}.
- */
-@MethodsReturnNonnullByDefault
-@ParametersAreNonnullByDefault
-package org.togetherjava.tjbot.routines;
-
-import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;
